Fix deadlock condition
This is a quick on-spot fix, would be nice to actually rework the
overall implementation to avoid this hackish solution.

The problem is that STI.Save (for saving artifacts) calls RunContainer
with opts.OnStart set to a function that reads from the read half of a
pipe, while the write half is set as the container stdout. Docker won't
close the container's output and error streams upon termination
(expected behavior), so OnStart will never terminate, it will block
forever waiting for more data to be read.

With OnStart blocked, RunContainer is blocked, and STI.Save never
returns, thus the deferred call to close the pipe halves is never
executed, thus the deadlock.