Added checks for existing content is git repo
This fixes an error when the git repo is empty, e.g. the user only ran
`git init`.  If the repo is empty, we use system copy commands instead
of git commands, and warn the user.

Also added errors for if there is a .git directory in the project, but
a Git binary is not installed.  Changed various function headers to throw
these errors.